ref(feedback): move trace section to below replay and add section title/icon
